refactored the heap and beam frontiers
They should be much faster (maybe even working). I specifically focused on the case where a node is visited that has a a state already in the frontier but a lower cost. Originally I would simply add the new node and ignore the old one, then I began to remove the old one but my process involved multiple traversals through the heap, first I checked if there was a node with the same state, then I checked if there was a node with the same state and higher cost. Now only one  traversal is done. Unfortunately, now the HeapFrontier cannot be simply swapped in place like StackFrontier and QueueFrontier can. BeamFrontier will also require a slightly different looking agent function.